refactor: Split up monolithic `roles` package into `forwarder`, `mounter` etc.
