+++ pkg/roles/registry/events.go
@@ -1,0 +1,8 @@
+package registry
+
+type CustomEventType byte
+
+const (
+	EventCustomAllDevicesSent    = CustomEventType(0)
+	EventCustomTransferAuthority = CustomEventType(1)
+)
