+++ pkg/roles/registry/migrate_to.go
@@ -1,120 +1,20 @@
-package roles
+package registry
 
 import (
 	"context"
 	"errors"
-	"fmt"
 	"io"
-	"os"
 	"sync/atomic"
-	"time"
 
-	iutils "github.com/loopholelabs/drafter/internal/utils"
+	"github.com/loopholelabs/drafter/internal/utils"
 	"github.com/loopholelabs/goroutine-manager/pkg/manager"
 	"github.com/loopholelabs/silo/pkg/storage"
-	"github.com/loopholelabs/silo/pkg/storage/blocks"
-	"github.com/loopholelabs/silo/pkg/storage/config"
-	"github.com/loopholelabs/silo/pkg/storage/device"
-	"github.com/loopholelabs/silo/pkg/storage/dirtytracker"
 	"github.com/loopholelabs/silo/pkg/storage/migrator"
-	"github.com/loopholelabs/silo/pkg/storage/modules"
 	"github.com/loopholelabs/silo/pkg/storage/protocol"
 	"github.com/loopholelabs/silo/pkg/storage/protocol/packets"
-	"github.com/loopholelabs/silo/pkg/storage/volatilitymonitor"
 )
 
-var (
-	ErrCouldNotGetInputDeviceStatistics = errors.New("could not get input device statistics")
-	ErrCouldNotCreateNewDevice          = errors.New("could not create new device")
-	ErrCouldNotSendDeviceInfo           = errors.New("could not send device info")
-	ErrCouldNotSendEvent                = errors.New("could not send event")
-	ErrCouldNotMigrate                  = errors.New("could not migrate")
-	ErrCouldNotContinueWithMigration    = errors.New("could not continue with migration")
-)
-
-type RegistryDevice struct {
-	Name      string `json:"name"`
-	Input     string `json:"input"`
-	BlockSize uint32 `json:"blockSize"`
-}
-
-type OpenedRegistryDevice struct {
-	RegistryDevice RegistryDevice
-
-	storage     *modules.Lockable
-	orderer     *blocks.PriorityBlockOrder
-	totalBlocks int
-	dirtyRemote *dirtytracker.DirtyTrackerRemote
-}
-
-type OpenDevicesHooks struct {
-	OnDeviceOpened func(deviceID uint32, name string)
-}
-
-func OpenDevices(
-	devices []RegistryDevice,
-
-	hooks OpenDevicesHooks,
-) ([]OpenedRegistryDevice, []func() error, error) {
-	openedDevices, deferFuncs, err := iutils.ConcurrentMap(
-		devices,
-		func(index int, input RegistryDevice, output *OpenedRegistryDevice, addDefer func(deferFunc func() error)) error {
-			output.RegistryDevice = input
-
-			stat, err := os.Stat(input.Input)
-			if err != nil {
-				return errors.Join(ErrCouldNotGetInputDeviceStatistics, err)
-			}
-
-			src, _, err := device.NewDevice(&config.DeviceSchema{
-				Name:      input.Name,
-				System:    "file",
-				Location:  input.Input,
-				Size:      fmt.Sprintf("%v", stat.Size()),
-				BlockSize: fmt.Sprintf("%v", input.BlockSize),
-				Expose:    false,
-			})
-			if err != nil {
-				return errors.Join(ErrCouldNotCreateNewDevice, err)
-			}
-			addDefer(src.Close)
-
-			dirtyLocal, dirtyRemote := dirtytracker.NewDirtyTracker(src, int(input.BlockSize))
-			output.dirtyRemote = dirtyRemote
-			monitor := volatilitymonitor.NewVolatilityMonitor(dirtyLocal, int(input.BlockSize), 10*time.Second)
-
-			storage := modules.NewLockable(monitor)
-			output.storage = storage
-			addDefer(func() error {
-				storage.Unlock()
-
-				return nil
-			})
-
-			totalBlocks := (int(storage.Size()) + int(input.BlockSize) - 1) / int(input.BlockSize)
-			output.totalBlocks = totalBlocks
-
-			orderer := blocks.NewPriorityBlockOrder(totalBlocks, monitor)
-			output.orderer = orderer
-			orderer.AddAll()
-
-			if hook := hooks.OnDeviceOpened; hook != nil {
-				hook(uint32(index), input.Name)
-			}
-
-			return nil
-		},
-	)
-
-	defers := []func() error{}
-	for _, deferFuncs := range deferFuncs {
-		defers = append(defers, deferFuncs...)
-	}
-
-	return openedDevices, defers, err
-}
-
-type MigrateDevicesHooks struct {
+type MigrateToHooks struct {
 	OnDeviceSent               func(deviceID uint32)
 	OnDeviceAuthoritySent      func(deviceID uint32)
 	OnDeviceMigrationProgress  func(deviceID uint32, ready int, total int)
@@ -124,7 +24,7 @@ type MigrateDevicesHooks struct {
 	OnAllMigrationsCompleted func()
 }
 
-func MigrateOpenedDevices(
+func MigrateTo(
 	ctx context.Context,
 
 	openedDevices []OpenedRegistryDevice,
@@ -133,7 +33,7 @@ func MigrateOpenedDevices(
 	readers []io.Reader,
 	writers []io.Writer,
 
-	hooks MigrateDevicesHooks,
+	hooks MigrateToHooks,
 ) (errs error) {
 	goroutineManager := manager.NewGoroutineManager(
 		ctx,
@@ -159,7 +59,7 @@ func MigrateOpenedDevices(
 
 	var devicesLeftToSend atomic.Int32
 
-	_, deferFuncs, err := iutils.ConcurrentMap(
+	_, deferFuncs, err := utils.ConcurrentMap(
 		openedDevices,
 		func(index int, input OpenedRegistryDevice, _ *struct{}, _ func(deferFunc func() error)) error {
 			to := protocol.NewToProtocol(input.storage.Size(), uint32(index), pro)
